created util for creating the drag handle element
